requirements.txt
pandas>=1.5.0
numpy>=1.20.0
matplotlib>=3.5.0
Pillow>=9.0.0
requests>=2.25.0
configparser>=5.0.0
networkx>=3.0
python-dotenv>=1.0.0
requests-cache>=1.0.0
tenacity>=8.0.0
joblib>=1.0.0